+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasetop.py
@@ -78,5 +78,5 @@ class RetrieveDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-07-29T10:30:09.739Z")
+    updated: Optional[datetime] = parse_datetime("2025-07-29T14:23:48.765Z")
     r"""The date and time the resource was last updated"""